# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhoto API
  version: 1.0.0
  description: WASAPhoto API for our university course on web and software architecture
tags:
  - name: login
  - name: user
  - name: comment
  - name: like
  - name: photo
  
security:
  - BearerAuth: []
  
  
paths:
  /session:
    post:
      tags: [login]
      summary: Logs in the user
      description: |
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: User's name
                  minLength: 3
                  maxLength: 50
                  pattern: "^[a-zA-Z0-9_]+$"  # Add a pattern for every string property
              description: General user object
                  
          
        required: true
      responses:
        '201':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'
                
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/ServerError" }
                

  /users/:
    post:
      tags: [user]
      summary: Create User
      description: Create a new user.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'

      responses:
        '201':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }

                
                
    
  /users/{username}/banned/:
    parameters:
        - name: username
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/username'
    post:
      tags: [user]
      summary: Ban User
      description: Ban a user.
      operationId: banUser
      responses:
        '201':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'
  
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
                
    delete:
      tags: [user]
      summary: Unban User
      description: Unban a user.
      operationId: unbanUser
  
      responses:
        '200':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'
  
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
                
  /stream:
    get:
      tags: [user]
      summary: Get My Stream
      description: Retrieve the user's stream.
      operationId: getMyStream
      
  
      responses:
        '200':
          description: User's stream retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  stream:
                    type: array
                    items:
                      $ref: '#/components/schemas/PhotoDetails'
                    minItems: 0
                    maxItems: 100
                    description: a photo detail object in the users stream
                description: The entire stream of photos
                
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
  /users/{username}:
    parameters:
        - name: username
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/username'
    patch:
      tags: [user]
      summary: Set My User Name
      description: Set the user's username.
      operationId: setMyUserName
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'

      responses:
        '200':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
  
    get:
      tags: [user]
      summary: Get User Profile
      description: Get the profile of a user.
      operationId: getUserProfile
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
                
                
                
  /users/{username}/followers/:
    parameters:
          - name: username
            in: path
            required: true
            schema:
              $ref: '#/components/schemas/username'
            description: the id of the user
    post:
      tags: [user]
      summary: Follow User
      description: Follow a user.
      operationId: followUser

      responses:
        '201':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
    delete:
      tags: [user]
      summary: Unfollow User
      description: Unfollow a user.
      operationId: unfollowUser
      responses:
        '200':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }

  /photos/{photoId}/comments/:
    parameters:
        - name: photoId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/photo_id'
    post:
      tags: [comment]
      summary: Comment on Photo
      description: Comment on a photo.
      operationId: commentPhoto
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
      responses:
        '201':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
    get:
      tags: [comment]
      summary: Get comments
      description: Get all comments of photo
      operationId: getComments
      responses:
        '200':
          description: Successfully getting the comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  stream:
                    type: array
                    items:
                      $ref: '#/components/schemas/PhotoDetails'
                    minItems: 0
                    maxItems: 100
                    description: a photo comment object for the photo
                description: The entire comments of the photo

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }

  /comments/{commentId}:
    delete:
      tags: [comment]
      summary: Remove Comment from Photo
      description: Remove a comment from a photo.
      operationId: uncommentPhoto
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
            minLength: 10
            maxLength: 20
            pattern: "^[a-zA-Z0-9_]+$"
            description: The comment ID

      responses:
        '200':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }

  /photos/{photoId}/likes/:
    parameters:
        - name: photoId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/photo_id'
    post:
      tags: [like, photo]
      summary: Like Photo
      description: Like a photo.
      operationId: likePhoto


      responses:
        '201':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
                
    delete:
      tags: [like]
      summary: Unlike Photo
      description: Unlike a photo.
      operationId: unlikePhoto
      responses:
        '201':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }



    

  /photos/:
    post:
      tags: [photo]
      summary: Upload Photo
      description: Upload a photo.
      operationId: uploadPhoto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Photo'
      responses:
        '201':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'
                

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }

  /photos/{photoId}:
    delete:
      tags: [photo]
      summary: Delete Photo
      description: Delete a photo.
      operationId: deletePhoto
      parameters:
        - name: photoId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/photo_id'
      responses:
        '200':
          description: action successful
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Success'

        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }

components:
  responses:
    BadRequest:
      description: Error code 400
    ServerError:
      description: Error code 500
    Unauthorized:
      description: Error code 401 
  schemas:
    Success:
      type: string
      minLength: 1
      maxLength: 20
      pattern: '^[a-zA-Z0-9]{10,20}$'
      description: A message saying everything worked out
    Comment:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '^[a-zA-Z0-9]{10,20}$'
          description: COntent of the comment
        commenter:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '^[a-zA-Z0-9]{10,20}$'
          description: username of the commenter
        photo_id:
          type: string
          minLength: 10
          maxLength: 20
          pattern: '^[a-zA-Z0-9_]+$'
          description: A unique photo identifier
        commentId:  
          type: string
          minLength: 10
          maxLength: 20
          pattern: "^[a-zA-Z0-9_]+$"
          description: The comment ID
      description: The overall comment object     
    
    User:
      type: object
      properties:
        username:
          type: string
          minLength: 4
          maxLength: 30
          pattern: "^[a-zA-Z0-9_]+$"
          description: User's username
        Following:
          type: integer
          minimum: 0
          maximum: 100000
          description: The number of followers the user has
        Followers:
          type: integer
          minimum: 0
          description: The number of people the user is following
        PhotosCount:
          type: integer
          minimum: 0
          maximum: 10000
          description: The number of images the user has posted
        Photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
          minItems: 0
          maxItems: 100
          description: List of photos posted by the user
      description: The entire user with all its attributes
          
    Photo:
      type: object
      properties:
        username:
          type: string
          minLength: 10
          maxLength: 20
          pattern: '^[a-zA-Z0-9_]+$'
          description: A unique username of the poster
        photo_id:
          type: string
          minLength: 10
          maxLength: 20
          pattern: '^[a-zA-Z0-9_]+$'
          description: A unique photo identifier
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the photo was posted
          minLength: 3
          maxLength: 10
        image_data:
          type: string
          description: Base64-encoded image data
          minLength: 10
          maxLength: 500000
          pattern: '^[0-1]+$'
          example: 'p1L8TP6uvt0bCmksSguCwSQ=='
      required:
        - image_data
      description: the photo object
    
    photo_id:
      type: string
      minLength: 10
      maxLength: 20
      pattern: '^[a-zA-Z0-9_]+$'
      description: A unique photo identifier

      
    username:
          type: string
          minLength: 10
          maxLength: 20
          pattern: "^[a-zA-Z0-9_]+$"
          description: A unique user identifier
    PhotoDetails:
      type: object
      properties:
        username:
          type: string
          description: Username of the user who posted the photo
          minLength: 10
          maxLength: 20
        photo_id:
          type: string
          minLength: 10
          maxLength: 20
          pattern: '^[a-zA-Z0-9_]+$'
          description: A unique photo identifier
        imageData:
          type: string
          format: byte
          description: Base64-encoded image data
          minLength: 10
          maxLength: 20
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the photo was uploaded
          minLength: 10
          maxLength: 20
        likesCount:
          type: integer
          description: Number of likes on the photo
        commentsCount:
          type: integer
          description: Number of comments on the photo
      description: An information schema about a photo which doesnt have to be stored
      
    
      
          
          

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 