# service/api/createUser.go
package api

import (
	"encoding/json"
	"fmt"
	"git.sapienzaapps.it/fantasticcoffee/fantastic-coffee-decaffeinated/service/api/reqcontext"
	"git.sapienzaapps.it/fantasticcoffee/fantastic-coffee-decaffeinated/service/database"
	"github.com/julienschmidt/httprouter"
	"net/http"
	// Import your database package here
)

// Assuming User struct is defined in your database package
// import your database package and refer to it as dbpkg

func createUser(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {

	// Decode the request body into a User struct
	db := ctx.Database
	var user database.User

	err := json.NewDecoder(r.Body).Decode(&user)
	if err != nil {
		w.WriteHeader(http.StatusBadRequest) // Sets the status code only
    	return
	}
	r.Body.Close()

	// Call AddUser method of the database object
	ctx.Logger.Info("Made it to the adduser call")

	err, token := db.AddUser(&user)

	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	w.WriteHeader(http.StatusCreated)
	w.Header().Set("Content-Type", "text/plain")
	ctx.Logger.Info("User created successfully")
	fmt.Fprint(w, token)
}